+++ pysradb/search.py
@@ -2,19 +2,20 @@
 """
 import os
 import re
-import requests
 import sys
 import time
 import urllib
-import pandas as pd
 import xml.etree.ElementTree as Et
 from json import JSONDecodeError
+
+import pandas as pd
+import requests
 from tqdm import tqdm
 
 from .exceptions import IncorrectFieldException
 from .exceptions import MissingQueryException
-from .utils import scientific_name_to_taxid
 from .utils import requests_3_retries
+from .utils import scientific_name_to_taxid
 
 SEARCH_REQUEST_TIMEOUT = 20
 SRA_SEARCH_GROUP_SIZE = 300